+++ gems/ast-merge-git/lib/ast/merge/git.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require 'ast/merge'
+require_relative 'git/version'
 require_relative 'git/corpus'
 require_relative 'git/local_benchmark'